# apps/api/openapi/tsc-api.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/agents/event/suggestions/{id}/approve": {
      "post": {
        "operationId": "EventAgentController_approveSuggestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventAgent"
        ]
      }
    },
    "/api/agents/event/suggestions/{id}/dismiss": {
      "post": {
        "operationId": "EventAgentController_dismissSuggestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventAgent"
        ]
      }
    },
    "/api/agents/events/{eventId}/insights": {
      "get": {
        "operationId": "EventAgentAliasController_getInsights",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventAgentAlias"
        ]
      }
    },
    "/api/agents/events/{eventId}/run": {
      "post": {
        "operationId": "EventAgentAliasController_run",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventAgentAlias"
        ]
      }
    },
    "/api/agents/events/suggestions/{id}/approve": {
      "post": {
        "operationId": "EventAgentAliasController_approveSuggestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventAgentAlias"
        ]
      }
    },
    "/api/agents/events/suggestions/{id}/dismiss": {
      "post": {
        "operationId": "EventAgentAliasController_dismissSuggestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventAgentAlias"
        ]
      }
    },
...
    "/api/events/{id}/intelligence/recommendations": {
      "get": {
        "operationId": "EventIntelligenceController_getRecommendations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligence"
        ]
      }
    },
    "/api/events/{id}/intelligence/predict": {
      "get": {
        "operationId": "EventIntelligenceController_getPredict",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligence"
        ]
      }
    },
    "/api/events/{id}/intelligence": {
      "get": {
        "operationId": "EventIntelligenceController_getIntelligence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligence"
        ]
      }
    },
    "/api/events/{id}/intelligence/refresh": {
      "post": {
        "operationId": "EventIntelligenceController_refreshIntelligence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligence"
        ]
      }
    },
...
    "/api/event-intelligence/{id}/intelligence/recommendations": {
      "get": {
        "operationId": "EventIntelligenceAliasController_getRecommendations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligenceAlias"
        ]
      }
    },
    "/api/event-intelligence/{id}/intelligence/predict": {
      "get": {
        "operationId": "EventIntelligenceAliasController_getPredict",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligenceAlias"
        ]
      }
    },
    "/api/event-intelligence/{id}/intelligence": {
      "get": {
        "operationId": "EventIntelligenceAliasController_getIntelligence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligenceAlias"
        ]
      }
    },
    "/api/event-intelligence/{id}/intelligence/refresh": {
      "post": {
        "operationId": "EventIntelligenceAliasController_refreshIntelligence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "EventIntelligenceAlias"
        ]
      }
    },
...
    "/api/contracts/{id}": {
      "get": {
        "operationId": "ContractController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Contract"
        ]
      }
    },
    "/api/contracts/{id}/sign": {
      "patch": {
        "operationId": "ContractController_sign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Contract"
        ]
      }
    },
    "/api/artists/{id}/contracts": {
      "get": {
        "operationId": "ArtistContractsController_listContracts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ArtistContracts"
        ]
      }
    },
    "/api/brands/{id}/contracts": {
      "get": {
        "operationId": "BrandContractsController_listContracts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BrandContracts"
        ]
      }
    },
...
    "/api/artists/{id}": {
      "get": {
        "operationId": "ArtistController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Artist"
        ]
      }
    },
...
    "/api/brands/{id}": {
      "get": {
        "operationId": "BrandController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Brand"
        ]
      },
      "patch": {
        "operationId": "BrandController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Brand"
        ]
      }
    },
...
  },
  "info": {
    "title": "TSC API",
    "description": "The Shakti Collective platform API — identity, community, marketplace, workspace, and intelligence services.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Local development"
    },
    {
      "url": "https://api.theshakticollective.in",
      "description": "Production"
    },
    {
      "url": "https://api-staging.theshakticollective.in",
      "description": "Staging"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Clerk session token (Authorization: Bearer <token>)"
      }
    },
...
  }
}